students/k3340/Gelm_Daniil/lab3/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
djoser==2.1.0
django-filter==23.3
drf-yasg==1.21.7
psycopg2-binary==2.9.11

